requirements.txt
Django==1.11.2
pytz==2017.2
django-oauth-toolkit
djangorestframework
psycopg2